Modified Healthmetrics to implement a NoOp version and modified tests to reflect the new NoOpHealthMetrics implementation
