+++ dd-trace-core/src/main/java/datadog/trace/core/CoreTracer.java
@@ -58,6 +58,7 @@ import datadog.trace.core.datastreams.DefaultDataStreamsCheckpointer;
 import datadog.trace.core.datastreams.StubDataStreamsCheckpointer;
 import datadog.trace.core.monitor.HealthMetrics;
 import datadog.trace.core.monitor.MonitoringImpl;
+import datadog.trace.core.monitor.TracerHealthMetrics;
 import datadog.trace.core.propagation.DatadogTags;
 import datadog.trace.core.propagation.ExtractedContext;
 import datadog.trace.core.propagation.HttpCodec;
@@ -482,8 +483,8 @@ public class CoreTracer implements AgentTracer.TracerAPI {
             : Monitoring.DISABLED;
     this.healthMetrics =
         config.isHealthMetricsEnabled()
-            ? new HealthMetrics(this.statsDClient)
-            : new HealthMetrics(StatsDClient.NO_OP);
+            ? new TracerHealthMetrics(this.statsDClient)
+            : HealthMetrics.NO_OP;
     this.healthMetrics.start();
     this.performanceMonitoring =
         config.isPerfMetricsEnabled()
