+++ dd-trace-core/src/main/java/datadog/trace/core/monitor/HealthMetrics.java
@@ -1,22 +1,8 @@
 package datadog.trace.core.monitor;
 
-import static datadog.trace.api.sampling.PrioritySampling.SAMPLER_DROP;
-import static datadog.trace.api.sampling.PrioritySampling.SAMPLER_KEEP;
-import static datadog.trace.api.sampling.PrioritySampling.USER_DROP;
-import static datadog.trace.api.sampling.PrioritySampling.USER_KEEP;
-import static java.util.concurrent.TimeUnit.SECONDS;
-
-import datadog.trace.api.StatsDClient;
-import datadog.trace.api.cache.RadixTreeCache;
 import datadog.trace.common.writer.RemoteApi;
 import datadog.trace.core.DDSpan;
-import datadog.trace.util.AgentTaskScheduler;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.IntFunction;
-import org.jctools.counters.CountersFactory;
-import org.jctools.counters.FixedSizeStripedLongCounter;
 
 /**
  * Callback for monitoring the health of the DDAgentWriter. Provides hooks for major lifecycle
@@ -30,265 +16,44 @@ import org.jctools.counters.FixedSizeStripedLongCounter;
  *   <li>sending to agent
  * </ul>
  */
-public class HealthMetrics implements AutoCloseable {
+public interface HealthMetrics extends AutoCloseable {
+  HealthMetrics NO_OP = new NoOpHealthMetrics();
 
-  private static final IntFunction<String[]> STATUS_TAGS =
-      httpStatus -> new String[] {"status:" + httpStatus};
+  void start();
 
-  private static final String[] NO_TAGS = new String[0];
-  private final RadixTreeCache<String[]> statusTagsCache =
-      new RadixTreeCache<>(16, 32, STATUS_TAGS, 200, 400);
+  void onStart(final int queueCapacity);
 
-  private final AtomicBoolean started = new AtomicBoolean(false);
-  private volatile AgentTaskScheduler.Scheduled<HealthMetrics> cancellation;
+  void onShutdown(final boolean flushSuccess);
 
-  private final FixedSizeStripedLongCounter userDropEnqueuedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter userKeepEnqueuedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter samplerDropEnqueuedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter samplerKeepEnqueuedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter unsetPriorityEnqueuedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
+  void onPublish(final List<DDSpan> trace, final int samplingPriority);
 
-  private final FixedSizeStripedLongCounter userDropDroppedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter userKeepDroppedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter samplerDropDroppedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter samplerKeepDroppedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter unsetPriorityDroppedTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
+  void onFailedPublish(final int samplingPriority);
 
-  private final FixedSizeStripedLongCounter enqueuedSpans =
-      CountersFactory.createFixedSizeStripedCounter(8);
+  void onPartialPublish(final int numberOfDroppedSpans);
 
-  private final FixedSizeStripedLongCounter createdTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
+  void onScheduleFlush(final boolean previousIncomplete);
 
-  private final FixedSizeStripedLongCounter createdSpans =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter sampledSpans =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter manualTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter cancelledContinuations =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter finishedContinuations =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter partialTraces =
-      CountersFactory.createFixedSizeStripedCounter(8);
-  private final FixedSizeStripedLongCounter droppedSpans =
-      CountersFactory.createFixedSizeStripedCounter(8);
+  void onFlush(final boolean early);
 
-  private final StatsDClient statsd;
-  private final long interval;
-  private final TimeUnit units;
+  void onPartialFlush(final int sizeInBytes);
 
-  public void start() {
-    if (started.compareAndSet(false, true)) {
-      cancellation =
-          AgentTaskScheduler.INSTANCE.scheduleAtFixedRate(
-              new Flush(), this, interval, interval, units);
-    }
-  }
+  void onSerialize(final int serializedSizeInBytes);
 
-  public HealthMetrics(final StatsDClient statsd) {
-    this(statsd, 30, SECONDS);
-  }
+  void onFailedSerialize(final List<DDSpan> trace, final Throwable optionalCause);
 
-  public HealthMetrics(final StatsDClient statsd, long interval, TimeUnit units) {
-    this.statsd = statsd;
-    this.interval = interval;
-    this.units = units;
-  }
+  void onCreateSpan();
 
-  public void onStart(final int queueCapacity) {
-    statsd.gauge("queue.max_length", queueCapacity, NO_TAGS);
-  }
+  void onCreateTrace();
 
-  public void onShutdown(final boolean flushSuccess) {}
+  void onCreateManualTrace();
 
-  public void onPublish(final List<DDSpan> trace, final int samplingPriority) {
-    switch (samplingPriority) {
-      case USER_DROP:
-        userDropEnqueuedTraces.inc();
-        break;
-      case USER_KEEP:
-        userKeepEnqueuedTraces.inc();
-        break;
-      case SAMPLER_DROP:
-        samplerDropEnqueuedTraces.inc();
-        break;
-      case SAMPLER_KEEP:
-        samplerKeepEnqueuedTraces.inc();
-        break;
-      default:
-        unsetPriorityEnqueuedTraces.inc();
-    }
-    enqueuedSpans.inc(trace.size());
-  }
+  void onCancelContinuation();
 
-  public void onFailedPublish(final int samplingPriority) {
-    switch (samplingPriority) {
-      case USER_DROP:
-        userDropDroppedTraces.inc();
-        break;
-      case USER_KEEP:
-        userKeepDroppedTraces.inc();
-        break;
-      case SAMPLER_DROP:
-        samplerDropDroppedTraces.inc();
-        break;
-      case SAMPLER_KEEP:
-        samplerKeepDroppedTraces.inc();
-        break;
-      default:
-        unsetPriorityDroppedTraces.inc();
-    }
-  }
+  void onFinishContinuation();
 
-  public void onPartialPublish(final int numberOfDroppedSpans) {
-    partialTraces.inc();
-    droppedSpans.inc(numberOfDroppedSpans);
-  }
+  void onSend(final int traceCount, final int sizeInBytes, final RemoteApi.Response response);
 
-  public void onScheduleFlush(final boolean previousIncomplete) {
-    // not recorded
-  }
+  void onFailedSend(final int traceCount, final int sizeInBytes, final RemoteApi.Response response);
 
-  public void onFlush(final boolean early) {}
-
-  public void onPartialFlush(final int sizeInBytes) {
-    statsd.count("span.flushed.partial", sizeInBytes, NO_TAGS);
-  }
-
-  public void onSerialize(final int serializedSizeInBytes) {
-    // DQH - Because of Java tracer's 2 phase acceptance and serialization scheme, this doesn't
-    // map precisely
-    statsd.count("queue.enqueued.bytes", serializedSizeInBytes, NO_TAGS);
-  }
-
-  public void onFailedSerialize(final List<DDSpan> trace, final Throwable optionalCause) {
-    // TODO - DQH - make a new stat for serialization failure -- or maybe count this towards
-    // api.errors???
-  }
-
-  public void onCreateSpan() {
-    createdSpans.inc();
-  }
-
-  public void onCreateTrace() {
-    createdTraces.inc();
-  }
-
-  public void onCreateManualTrace() {
-    manualTraces.inc();
-  }
-
-  public void onCancelContinuation() {
-    cancelledContinuations.inc();
-  }
-
-  public void onFinishContinuation() {
-    finishedContinuations.inc();
-  }
-
-  public void onSend(
-      final int traceCount, final int sizeInBytes, final RemoteApi.Response response) {
-    onSendAttempt(traceCount, sizeInBytes, response);
-  }
-
-  public void onFailedSend(
-      final int traceCount, final int sizeInBytes, final RemoteApi.Response response) {
-    onSendAttempt(traceCount, sizeInBytes, response);
-  }
-
-  private void onSendAttempt(
-      final int traceCount, final int sizeInBytes, final RemoteApi.Response response) {
-    statsd.incrementCounter("api.requests.total", NO_TAGS);
-    statsd.count("flush.traces.total", traceCount, NO_TAGS);
-    // TODO: missing queue.spans (# of spans being sent)
-    statsd.count("flush.bytes.total", sizeInBytes, NO_TAGS);
-
-    if (response.exception() != null) {
-      // covers communication errors -- both not receiving a response or
-      // receiving malformed response (even when otherwise successful)
-      statsd.incrementCounter("api.errors.total", NO_TAGS);
-    }
-
-    if (response.status() != null) {
-      statsd.incrementCounter("api.responses.total", statusTagsCache.get(response.status()));
-    }
-  }
-
-  @Override
-  public void close() {
-    if (null != cancellation) {
-      cancellation.cancel();
-    }
-  }
-
-  private static class Flush implements AgentTaskScheduler.Task<HealthMetrics> {
-
-    private static final String[] USER_DROP_TAG = new String[] {"priority:user_drop"};
-    private static final String[] USER_KEEP_TAG = new String[] {"priority:user_keep"};
-    private static final String[] SAMPLER_DROP_TAG = new String[] {"priority:sampler_drop"};
-    private static final String[] SAMPLER_KEEP_TAG = new String[] {"priority:sampler_keep"};
-    private static final String[] UNSET_TAG = new String[] {"priority:unset"};
-
-    @Override
-    public void run(HealthMetrics target) {
-      reportIfChanged(
-          target.statsd, "queue.enqueued.traces", target.userDropEnqueuedTraces, USER_DROP_TAG);
-      reportIfChanged(
-          target.statsd, "queue.enqueued.traces", target.userKeepEnqueuedTraces, USER_KEEP_TAG);
-      reportIfChanged(
-          target.statsd,
-          "queue.enqueued.traces",
-          target.samplerDropEnqueuedTraces,
-          SAMPLER_DROP_TAG);
-      reportIfChanged(
-          target.statsd,
-          "queue.enqueued.traces",
-          target.samplerKeepEnqueuedTraces,
-          SAMPLER_KEEP_TAG);
-      reportIfChanged(
-          target.statsd, "queue.enqueued.traces", target.unsetPriorityEnqueuedTraces, UNSET_TAG);
-      reportIfChanged(
-          target.statsd, "queue.dropped.traces", target.userDropDroppedTraces, USER_DROP_TAG);
-      reportIfChanged(
-          target.statsd, "queue.dropped.traces", target.userKeepDroppedTraces, USER_KEEP_TAG);
-      reportIfChanged(
-          target.statsd, "queue.dropped.traces", target.samplerDropDroppedTraces, SAMPLER_DROP_TAG);
-      reportIfChanged(
-          target.statsd, "queue.dropped.traces", target.samplerKeepDroppedTraces, SAMPLER_KEEP_TAG);
-      reportIfChanged(
-          target.statsd, "queue.dropped.traces", target.unsetPriorityDroppedTraces, UNSET_TAG);
-      reportIfChanged(target.statsd, "queue.enqueued.spans", target.enqueuedSpans, NO_TAGS);
-      reportIfChanged(target.statsd, "trace.pending.created", target.createdTraces, NO_TAGS);
-      reportIfChanged(target.statsd, "span.pending.created", target.createdSpans, NO_TAGS);
-      reportIfChanged(
-          target.statsd, "span.continuations.canceled", target.cancelledContinuations, NO_TAGS);
-      reportIfChanged(
-          target.statsd, "span.continuations.finished", target.finishedContinuations, NO_TAGS);
-      reportIfChanged(target.statsd, "queue.partial.traces", target.partialTraces, NO_TAGS);
-      reportIfChanged(target.statsd, "queue.dropped.spans", target.droppedSpans, NO_TAGS);
-    }
-
-    private void reportIfChanged(
-        StatsDClient statsDClient,
-        String aspect,
-        FixedSizeStripedLongCounter counter,
-        String[] tags) {
-      long count = counter.getAndReset();
-      if (count > 0) {
-        statsDClient.count(aspect, count, tags);
-      }
-    }
-  }
+  void close();
 }
